[Jun235] add [member_service_test]

+++ src/main/java/com/nhnacademy/minidoorayaccountapi/member_authority/repository/MemberAuthorityRepository.java
@@ -3,5 +3,5 @@ package com.nhnacademy.minidoorayaccountapi.member_authority.repository;
 import com.nhnacademy.minidoorayaccountapi.member_authority.entity.MemberAuthority;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberAuthorityRepository extends JpaRepository<MemberAuthority, Long> {
+public interface MemberAuthorityRepository extends JpaRepository<MemberAuthority, Integer> {
 }
